refactor(RacingController): 기존 2개로 나눠져있던 controller 통합
