+++ src/main/java/racingcar/view/outputview/OutputView.java
@@ -1,16 +1,15 @@
 package racingcar.view.outputview;
 
+import java.util.List;
 import java.util.StringJoiner;
-import java.util.stream.Collectors;
+import racingcar.controller.CarResponse;
+import racingcar.controller.TrackResponse;
 import racingcar.exception.CustomException;
-import racingcar.model.car.Car;
-import racingcar.model.car.Cars;
 
 public class OutputView {
     public static final String POSITION_CAR_FORMAT_SYMBOL = "-";
     public static final String POSITION_CAR_STATE_FORMAT = "%s : %s";
     private static final String WINNER_MESSAGE = "%s가 최종 우승했습니다.";
-    private static final String CAR_SEPARATOR = ", ";
     private static final String LINE_BREAK = "\n";
 
     private OutputView() {
@@ -20,23 +19,18 @@ public class OutputView {
         System.out.println(exception.getMessage());
     }
 
-    public static void printCurrentCarsPosition(final Cars cars) {
+    public static void printCurrentCarsPosition(final TrackResponse trackResponse) {
         StringJoiner stringJoiner = new StringJoiner(LINE_BREAK);
 
-        cars.getCarsCurrentInfo().forEach(car -> {
-            stringJoiner.add(String.format(POSITION_CAR_STATE_FORMAT,
-                    car.getCarName(),
-                    POSITION_CAR_FORMAT_SYMBOL.repeat(car.getPosition())));
-        });
+        List<CarResponse> cars = trackResponse.getRacingCars();
+        cars.forEach(car -> stringJoiner.add(String.format(POSITION_CAR_STATE_FORMAT,
+                car.getName(),
+                POSITION_CAR_FORMAT_SYMBOL.repeat(car.getPosition()))));
         System.out.println(stringJoiner + LINE_BREAK);
     }
 
 
-    public static void printWinnerCars(final Cars cars) {
-        String winnerCarsFormat = cars.getWinnerCars().stream()
-                .map(Car::getCarName)
-                .collect(Collectors.joining(CAR_SEPARATOR));
-
-        System.out.println(String.format(WINNER_MESSAGE, winnerCarsFormat));
+    public static void printWinnerCars(final TrackResponse trackResponse) {
+        System.out.printf((WINNER_MESSAGE) + "%n", trackResponse.getWinners());
     }
 }
